EditDialog: refactor PresetSelect - move to folder
